pyproject.toml
[tool.poetry]
name = "font-recognition-ja"
version = "0.1.0"
description = ""
authors = ["masashi kishimoto <drehbleistift@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.11.0"
torchvision = "^0.12.0"
numpy = "^1.22.3"
albumentations = "^1.1.0"
Pillow = "^9.0.1"
imutils = "^0.5.4"
scikit-learn = "^1.0.2"
tqdm = "^4.63.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
jupyter = "^1.0.0"
jupyterlab = "^3.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
